change DisplayView and subclasses so that setPosition() works in longs
git-svn-id: http://code.imagej.net/svn/imagej/trunk@3171 be23795f-a60e-40d4-ba19-055798872e57

+++ core/display/src/main/java/imagej/display/DisplayView.java
@@ -69,7 +69,7 @@ public interface DisplayView {
 	long getPosition(final int dim);
 
 	/** Sets the position of the given dimensional axis. */
-	void setPosition(final int value, final int dim);
+	void setPosition(final long value, final int dim);
 
 	/** Gets the view's ideal width in pixels. */
 	int getPreferredWidth();
